# go/api_default.go
/*
 * API POC Server
 *
 * This is a proof of concept OpenAPI server
 *
 * API version: 0.0.1
 * Generated by: OpenAPI Generator (https://openapi-generator.tech)
 */

package openapi

import (
	"encoding/json"
	"fmt"
	"log"
	"net/http"
)

func AddSite(w http.ResponseWriter, r *http.Request) {
	var si Site

	if r.Body == nil {
		w.WriteHeader(http.StatusBadRequest)
		_, err := w.Write([]byte("Please send a request body"))
		if err != nil {
			log.Print("Failed to send message")
		}
		return
	}
	err := json.NewDecoder(r.Body).Decode(&si)
	if err != nil {
		w.WriteHeader(http.StatusBadRequest)
		_, err := w.Write([]byte(err.Error()))
		if err != nil {
			log.Print("Failed to send message")
		}
		return
	}
	//if id.Operation == "ADD" {
	//	answer = id.A + id.B
	//} else if id.Operation == "SUBTRACT" {
	//	answer = id.A - id.B
	//} else if id.Operation == "MULTIPLY" {
	//	answer = id.A * id.B
	//} else {
	//	w.WriteHeader(http.StatusTeapot)
	//	_, err := w.Write([]byte(fmt.Sprintf("Invalid operator: %s", id.Operation)))
	//	if err != nil {
	//		log.Print("Failed to sent message")
	//	}
	//	return
	//}

	w.Header().Set("Content-Type", "application/json; charset=UTF-8")
	w.WriteHeader(http.StatusOK)
	result, err := json.Marshal(&Site{
		TunnelId: si.TunnelId,
		PublicIp: si.PublicIp,
		MgmtIp: si.MgmtIp,
		IpClass: si.IpClass,
		Hostname: si.Hostname,
		Cost: si.Cost,
		Email: si.Email,
		SnmpCommunity: si.SnmpCommunity,
		TunnelType: si.TunnelType,
		TopologyType: si.TopologyType,

	})
	if err != nil {
		fmt.Println(err)
		return
	}
	_, err = w.Write(result)
	if err != nil {
		log.Print("Failed to sent message")
	}
}