# go/model_site.go
/*
 * API POC Server
 *
 * This is a proof of concept OpenAPI server
 *
 * API version: 0.0.2
 * Generated by: OpenAPI Generator (https://openapi-generator.tech)
 */

package openapi

type Site struct {

	// The ID of the tunnel [Ignored for site creation]
	TunnelId int32 `json:"tunnelId,omitempty"`

	// The public IP address of the router
	PublicIp string `json:"public_ip,omitempty"`

	// The IP address used by the controller to talk to the router
	MgmtIp string `json:"mgmt_ip,omitempty"`

	// The IP Class of this connection (IPv4 or IPv6) [IPv4 default]
	IpClass int32 `json:"ip_class,omitempty"`

	// The public DNS name of the router (for dynamic DNS lookups)
	Hostname string `json:"hostname,omitempty"`

	// The DECnet routing cost to assign to this tunnel [20 default]
	Cost int32 `json:"cost,omitempty"`

	// The email to send notifications to
	Email string `json:"email,omitempty"`

	// A RW SNMP community on the router to make updates with
	SnmpCommunity string `json:"snmp_community,omitempty"`

	// The tunnel type for this tunnel. GRE is the only option currently. [ignored]
	TunnelType string `json:"tunnel_type,omitempty"`

	// The topology type of this endpoint. (mesh, hub or spoke) [mesh default]
	TopologyType string `json:"topology_type,omitempty"`
}