# openapiui/openapi.json
{
  "openapi" : "3.0.0",
  "info" : {
    "title" : "API POC Server",
    "description" : "This is a proof of concept OpenAPI server",
    "version" : "0.0.2"
  },
  "servers" : [ {
    "url" : "/"
  } ],
  "paths" : {
    "/site" : {
      "post" : {
        "operationId" : "AddSite",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/Site"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "returns the newly created site",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Site"
                }
              }
            }
          },
          "400" : {
            "description" : "Error (see message)"
          }
        },
        "x-amazon-apigateway-integration" : {
          "uri" : [ "", [ "arn:aws:apigateway:", "AWS::Region", ":lambda:path/2015-03-31/functions/", "TunnelManagerLambdaFunction.Arn", "/invocations" ] ],
          "responses" : {
            "default" : {
              "statusCode" : "200"
            }
          },
          "passthroughBehavior" : "when_no_match",
          "httpMethod" : "POST",
          "contentHandling" : "CONVERT_TO_TEXT",
          "type" : "aws_proxy"
        }
      }
    }
  },
  "components" : {
    "schemas" : {
      "Site" : {
        "title" : "Site Information",
        "type" : "object",
        "properties" : {
          "tunnelId" : {
            "type" : "integer",
            "description" : "The ID of the tunnel [Ignored for site creation]",
            "format" : "int32"
          },
          "public_ip" : {
            "pattern" : "^(?:[0-9]{1,3}\\.){3}[0-9]{1,3}$",
            "type" : "string",
            "description" : "The public IP address of the router",
            "format" : "ipv4"
          },
          "mgmt_ip" : {
            "pattern" : "^(?:[0-9]{1,3}\\.){3}[0-9]{1,3}$",
            "type" : "string",
            "description" : "The IP address used by the controller to talk to the router",
            "format" : "ipv4"
          },
          "ip_class" : {
            "type" : "integer",
            "description" : "The IP Class of this connection (IPv4 or IPv6) [IPv4 default]",
            "format" : "int32",
            "enum" : [ 4, 6 ]
          },
          "hostname" : {
            "type" : "string",
            "description" : "The public DNS name of the router (for dynamic DNS lookups)"
          },
          "cost" : {
            "maximum" : 63,
            "minimum" : 1,
            "type" : "integer",
            "description" : "The DECnet routing cost to assign to this tunnel [20 default]",
            "format" : "int32"
          },
          "email" : {
            "pattern" : "/^(([^<>()\\[\\]\\\\.,;:\\s@\"]+(\\.[^<>()\\[\\]\\\\.,;:\\s@\"]+)*)|(\".+\"))@((\\[[0-9]{1,3}\\.[0-9]{1,3}\\.[0-9]{1,3}\\.[0-9]{1,3}])|(([a-zA-Z\\-0-9]+\\.)+[a-zA-Z]{2,}))$/",
            "type" : "string",
            "description" : "The email to send notifications to",
            "format" : "email"
          },
          "snmp_community" : {
            "type" : "string",
            "description" : "A RW SNMP community on the router to make updates with"
          },
          "tunnel_type" : {
            "type" : "string",
            "description" : "The tunnel type for this tunnel. GRE is the only option currently. [ignored]",
            "enum" : [ "GRE", "IPSec" ]
          },
          "topology_type" : {
            "type" : "string",
            "description" : "The topology type of this endpoint. (mesh, hub or spoke) [mesh default]",
            "enum" : [ "mesh", "hub", "spoke" ]
          }
        },
        "example" : {
          "tunnelId" : 81,
          "public_ip" : "8.8.8.8",
          "mgmt_ip" : "172.16.0.12",
          "ip_class" : 4,
          "hostname" : "myrouter.dyndns.com",
          "cost" : 20,
          "email" : "nobody@nowhere.com",
          "snmp_community" : "letmein",
          "tunnel_type" : "GRE",
          "topology_type" : "mesh"
        }
      },
      "OutputData" : {
        "title" : "The results of the operation",
        "type" : "object",
        "properties" : {
          "result" : {
            "type" : "integer",
            "description" : "The Answer",
            "format" : "int64"
          }
        },
        "example" : {
          "result" : 7
        }
      }
    }
  }
}